+++ client/src/components/Plan/Plan.jsx
@@ -1,17 +1,23 @@
-import React from "react";
+import React, { useState } from "react";
 import GeolocationForm from "./GeolocationForm";
 import PlanCardList from "../../recycle/PlanCardList/PlansCardList";
 import PlansSections from "../../recycle/PlansSections/PlansSections";
 import style from "./Plan.module.css";
+import PlanForm from "./PlanForm";
 
 export default function Plan() {
+  const [showPlanForm, setShowPlanForm] = useState(false);
+  const handleSubmit = () => {
+    setShowPlanForm(true);
+  };
   return (
     <div className={style.container}>
-      <GeolocationForm />
-      <div className={style.plans}>
+      {showPlanForm && <PlanForm setShowPlanForm={setShowPlanForm} />}
+        {!showPlanForm && <GeolocationForm setShowPlanForm={setShowPlanForm}/>}
+      {!showPlanForm && <div className={style.plans}>
         <PlanCardList />
-      </div>
-      <PlansSections />
+      </div>}
+      {!showPlanForm && <PlansSections />}
     </div>
   );
 }
